Avoid non_fmt_panic warning
This warning is enabled by default in Rust 1.50.0, and gets triggered by
curly braces in the `assert!`: Rust 2021 will treat the string as
a format string, and thus try to process those curly braces. However,
Rust 2015 and 2018 don't do that, so escaping the braces as per std::fmt
convention doesn't work. That's why I switched the whole `assert!` to
a three-argument form.